+++ CHANGES.md
@@ -1,4 +1,5 @@
 ## 2.0.0 / 2019-11-25
+* fix processing of .json files, thanks @ahmader
 * use gettext/gettext to parse translation headers for 'PO-Revision-Date'
 
 ## 1.8.0 / 2019-11-24